# mandrel-command/frontend/.openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIDIS Command Backend API",
    "version": "1.0.0",
    "description": "RESTful API for AIDIS Command Backend with MCP Bridge Integration",
    "contact": {
      "name": "AIDIS Development Team",
      "email": "dev@aidis.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data payload"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation ID for tracking"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "version": {
                "type": "string",
                "example": "1.0.0"
              }
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "validation",
                  "authentication",
                  "authorization",
                  "not_found",
                  "internal",
                  "business"
                ],
                "description": "Error type for client handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details (e.g., validation errors)"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              }
            },
            "required": [
              "type",
              "message"
            ]
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "BaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Project name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "Project description"
          },
          "git_repo_url": {
            "type": "string",
            "format": "uri",
            "description": "Git repository URL"
          },
          "root_directory": {
            "type": "string",
            "maxLength": 255,
            "description": "Project root directory path"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Additional project metadata"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ContextEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "description": "Context body content"
              },
              "type": {
                "type": "string",
                "enum": [
                  "code",
                  "decision",
                  "research",
                  "issue",
                  "note",
                  "error",
                  "test"
                ],
                "description": "Context classification"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Associated tags"
              },
              "metadata": {
                "type": "object",
                "additionalProperties": true,
                "description": "Arbitrary metadata payload"
              },
              "session_id": {
                "type": "string",
                "format": "uuid",
                "description": "Linked session identifier"
              },
              "project_id": {
                "type": "string",
                "format": "uuid",
                "description": "Associated project identifier"
              },
              "relevance_score": {
                "type": "number",
                "description": "Semantic similarity score when applicable"
              }
            },
            "required": [
              "content",
              "type"
            ]
          }
        ]
      },
...
      "ContextSearchResponse": {
        "type": "object",
        "properties": {
          "contexts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextEntity"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "contexts",
          "total"
        ]
      },
...
      "DecisionEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "problem": {
                "type": "string"
              },
              "decision": {
                "type": "string"
              },
              "rationale": {
                "type": "string"
              },
              "alternatives": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "under_review",
                  "superseded",
                  "deprecated"
                ]
              },
              "outcomeStatus": {
                "type": "string",
                "enum": [
                  "unknown",
                  "successful",
                  "failed",
                  "mixed",
                  "too_early"
                ],
                "description": "Lifecycle outcome assessment"
              },
              "outcomeNotes": {
                "type": "string"
              },
              "lessonsLearned": {
                "type": "string"
              },
              "supersededBy": {
                "type": "string"
              },
              "supersededReason": {
                "type": "string"
              },
              "outcome": {
                "type": "string"
              },
              "lessons": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "project_id": {
                "type": "string",
                "format": "uuid"
              },
              "project_name": {
                "type": "string"
              },
              "created_by": {
                "type": "string"
              },
              "updated_by": {
                "type": "string"
              }
            },
            "required": [
              "title",
              "problem",
              "decision",
              "status"
            ]
          }
        ]
      },
      "DecisionSearchResponse": {
        "type": "object",
        "properties": {
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionEntity"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        },
        "required": [
          "decisions",
          "total"
        ]
      },
...
      "CreateDecisionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "problem": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "decision": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "rationale": {
            "type": "string",
            "maxLength": 5000
          },
          "alternatives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "under_review",
              "superseded",
              "deprecated"
            ],
            "default": "active"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "problem",
          "decision"
        ]
      },
...
      "NamingSearchResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamingEntry"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        },
        "required": [
          "entries",
          "total"
        ]
      },
...
      "ProjectListResponse": {
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectEntity"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "projects",
          "total"
        ]
      },
      "ProjectDetailResponse": {
        "type": "object",
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectEntity"
          }
        },
        "required": [
          "project"
        ]
      },
...
      "SessionListResponse": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionEntity"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "sessions",
          "total"
        ]
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/contexts": {
      "get": {
        "summary": "Search contexts with filters",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Free-text search query"
          },
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by project ID (falls back to X-Project-ID header)"
          },
          {
            "in": "query",
            "name": "session_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "code",
                "decision",
                "research",
                "issue",
                "note",
                "error",
                "test"
              ]
            }
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated tags filter"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contexts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContextSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/decisions": {
      "get": {
        "summary": "Search technical decisions",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Free-text search term to locate decisions"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "under_review",
                "superseded",
                "deprecated"
              ]
            },
            "description": "Filter decisions by lifecycle status"
          },
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Scope search to a specific project"
          },
          {
            "in": "query",
            "name": "created_by",
            "schema": {
              "type": "string"
            },
            "description": "Filter by author username"
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive lower bound for creation date"
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive upper bound for creation date"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decisions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DecisionSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Record a new technical decision",
        "tags": [
          "Decisions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Decision recorded successfully"
          }
        }
      }
    },
...
    "/naming": {
      "get": {
        "summary": "Search naming registry entries",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "variable",
                "function",
                "component",
                "class",
                "interface",
                "module",
                "file"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "deprecated",
                "conflicted",
                "pending"
              ]
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "created_by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entries retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/projects": {
      "get": {
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "asc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectListResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProjectDetailResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid project data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions": {
      "get": {
        "summary": "List sessions with optional filters",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "Max number of sessions to return (end-to-end pagination)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "description": "Number of sessions to skip",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessions returned",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionListResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}